feature/providing-ratatui-and-clap: Clippy fix
