+++ runner/src/tui/mod.rs
@@ -1,4 +1,67 @@
 pub mod event;
-pub mod tui;
 pub mod ui;
 pub mod update;
+
+use std::io;
+use std::panic;
+
+use anyhow::Result;
+
+use crossterm::terminal;
+use crossterm::terminal::EnterAlternateScreen;
+use crossterm::terminal::LeaveAlternateScreen;
+
+use crossterm::event::EnableMouseCapture;
+use crossterm::event::DisableMouseCapture;
+
+pub type Frame<'a> = ratatui::Frame<'a, ratatui::backend::CrosstermBackend<std::io::Stderr>>;
+pub type CrosstermTerminal = ratatui::Terminal<ratatui::backend::CrosstermBackend<std::io::Stderr>>;
+
+use crate::app::App;
+use crate::tui::event::EventHandler;
+
+pub struct Tui {
+    terminal: CrosstermTerminal,
+    pub events: EventHandler,
+}
+
+impl Tui {
+    pub fn new(terminal: CrosstermTerminal, events: EventHandler) -> Self {
+        Self { 
+            terminal, 
+            events
+        }
+    }
+
+    pub fn enter(&mut self) -> Result<()> {
+        terminal::enable_raw_mode()?;
+        crossterm::execute!(io::stderr(), EnterAlternateScreen, EnableMouseCapture)?;
+
+        let panic_hook = panic::take_hook();
+        panic::set_hook(Box::new(move |panic| {
+            Self::reset().expect("failed to reset the terminal");
+            panic_hook(panic);
+        }));
+
+        self.terminal.hide_cursor()?;
+        self.terminal.clear()?;
+        Ok(())
+    }
+
+    pub fn draw(&mut self, app: &mut App) -> Result<()> {
+        self.terminal.draw(|frame| ui::render(app, frame))?;
+        Ok(())
+    }
+
+    fn reset() -> Result<()> {
+        terminal::disable_raw_mode()?;
+        crossterm::execute!(io::stderr(), LeaveAlternateScreen, DisableMouseCapture)?;
+        Ok(())
+    }
+
+    pub fn exit(&mut self) -> Result<()> {
+        Self::reset()?;
+        self.terminal.show_cursor()?;
+        Ok(())
+    }
+}
